Changed characterList from an ArrayList to a List for easier type control.
Removed unnecessary comments.

Changed some "breaks" to "continue" to ensure other characters are checked instead of quitting.

Added a checkbox to see if you want to add initiative modifier to PCs or just input their overall initiative. (This is because a lot of players like to roll their own initiative and add respective modifiers.)

Added first attempt to parse for initiative modifier. Need to implement so it checks if you do it at the end of a characters name when first adding. Formats: +(int), -(int), (int). Will set whatever (int) is to the initiative modifier.

Added a "Change Initiative modifier" when rightclicking on a character. Will probably remove.

Added an Up/Down box to easily and quickly change value of initiative modifier.