Tried to fix up how some stuff looked by aligning the status effect form. Also added some more stuff for combat form.
I broke the Combat Form by trying to fix leaving and entering it causing it to reset the round number. I now want to the combatform to always be open instead of making a new one each time, but I will need to weigh the pros and cons. I might make a new branch for each approach.

Also added temp hp.